+++ src/useTimeout.spec.js
@@ -8,8 +8,9 @@ describe('useTimeout', () => {
   beforeEach(() => {
     cleanupHooks();
     cleanupReact();
-    sinon.restore();
   });
+
+  afterEach(sinon.restore);
 
   it('should be a function', () => {
     expect(useTimeout).to.be.a('function');