+++ src/utils/isAPISupported.js
@@ -1,0 +1,4 @@
+/**
+ * Exports a boolean value reporting whether the given API is supported or not
+ */
+export default (api) => (api in window);